# openapi.yaml
openapi: 3.0.0
info:
  title: API
  description: ""
  version: 1.0.0
servers:
- url: http://localhost:3000
- url: ""
paths:
...
  /api/user/login:
    post:
      tags:
      - User Tests
      summary: Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
            example: |-
              //{
              //    "email": "system@admin.com",
              //    "password": "password"

              //}

               {
                   "email": "testemail2@mail.com",
                   "password": "123456"

               }
      responses:
        "200":
          description: succesful
...
  /api/social/threads:
    get:
      tags:
      - Thread Tests
      - Pull tests
      summary: "search thread by title, owner, team, tags"
      parameters:
      - name: search
        in: query
        required: false
        schema:
          type: string
        example: bar
      - name: title
        in: query
        required: false
        schema:
          type: boolean
        example: "true"
      - name: author
        in: query
        required: false
        schema:
          type: boolean
        example: "false"
      - name: team
        in: query
        required: false
        schema:
          type: boolean
        example: "false"
      - name: tags
        in: query
        required: false
        schema:
          type: boolean
        example: "false"
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
        example: "Bearer {{accessToken}}"
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: succesful
    post:
      tags:
      - Thread Tests
      - Creation Tests
      summary: New General Thread
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
        example: "Bearer {{accessToken}}"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/social_threads_body"
            example:
              title: test
              text: test
              tags:
              - a
              - b
              - c
      responses:
        "200":
          description: succesful
    delete:
      tags:
      - Thread Tests
      - Delete tests
      summary: Delete thread
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
        example: "Bearer {{accessToken}}"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/social_threads_body_1"
            example:
              tid: 1
      responses:
        "200":
          description: succesful
...
  /api/social/moderation:
    post:
      tags:
      - moderation
      summary: Report Thread
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
        example: "Bearer {{accessToken}}"
      responses:
        "200":
          description: succesful
...
  /api/refresh:
    get:
      summary: Refresh
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
        example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjAsInVzZXJuYW1lIjoic3lzdGVtdXNlciIsInVzZXJFbWFpbCI6InN5c3RlbUBtYWlsLmNvbSIsImlhdCI6MTc3MjMzNzE0OCwiZXhwIjoxNzcyMzUxNTQ4fQ.qDwyUJ8gHehHRXm3mN5EXjmdDSPNTwNxy0ebrne3_vE
      responses:
        "200":
          description: succesful
      security:
      - BearerAuth: []
components:
  schemas:
...
    social_threads_body_1:
      type: object
      properties:
        tid:
          type: integer
...
    social_threads_body:
      type: object
      properties:
        title:
          type: string
        text:
          type: string
        tags:
          type: array
          items:
            type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer